google_apis: fix code for newer protobuf
This is to fix build error when we update protobuf in chromium build.

Bug: 40285204
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6169782
Commit-Queue: Colin Blundell <blundell@chromium.org>
Auto-Submit: Takuto Ikuta <tikuta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407784}

+++ google_apis/gcm/base/mcs_message.h
@@ -6,6 +6,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef GOOGLE_APIS_GCM_BASE_MCS_MESSAGE_H_
 #define GOOGLE_APIS_GCM_BASE_MCS_MESSAGE_H_
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include <memory>
@@ -47,7 +48,7 @@ class GCM_EXPORT MCSMessage {
 
   // Getters for serialization.
   uint8_t tag() const { return tag_; }
-  int size() const {return size_; }
+  size_t size() const { return size_; }
   std::string SerializeAsString() const;
 
   // Getter for accessing immutable probotuf fields.
@@ -79,7 +80,7 @@ class GCM_EXPORT MCSMessage {
 
   // These are cached separately to avoid having to recompute them.
   const uint8_t tag_;
-  const int size_;
+  const size_t size_;
 
   // The refcounted core, containing the protobuf memory.
   scoped_refptr<const Core> core_;
